move linechart reference line calculation into useLineChartLegend hook
