+++ common/hooks/useLineChartLegend.ts
@@ -15,8 +15,17 @@
 // along with this program.  If not, see <https://www.gnu.org/licenses/>.
 // =============================================================================
 
-import { filter, head, keys, map, mergeAll, pipe, reverse } from "ramda";
-import { useEffect, useState } from "react";
+import {
+  filter,
+  head,
+  keys,
+  map,
+  mergeAll,
+  pipe,
+  reduce,
+  reverse,
+} from "ramda";
+import { useEffect, useMemo, useState } from "react";
 
 import { LegendData } from "../components/DataViz/types";
 import { Datapoint } from "../types";
@@ -57,5 +66,29 @@ export const useLineChartLegend = (
 
     setLegendData(pipe(getLatestDatapoint, transformDataForLegend)(data));
   }, [data, colorDict, dimensions]);
-  return { legendData };
+
+  const referenceLineHeight = useMemo(
+    () =>
+      reduce(
+        (total: number, dimension: keyof Datapoint) => {
+          const reducedDatapoints = pipe(
+            filter(
+              (datapoint: Datapoint) => typeof datapoint[dimension] === "number"
+            ),
+            reduce(
+              (acc: number, datapoint: Datapoint): number =>
+                Number(datapoint[dimension]) > acc
+                  ? Number(datapoint[dimension])
+                  : acc,
+              0
+            )
+          )(data);
+          return reducedDatapoints > total ? reducedDatapoints : total;
+        },
+        0,
+        dimensions
+      ),
+    [dimensions]
+  );
+  return { legendData, referenceLineHeight };
 };
